add env variable configuration
breakup terraform deployment function

add database scoped env var key collision checking
